Removes double logging messages; simplifies statusUpdate function removing reqLogger parameter; adds context to log messages
The statusUpdate function in the account controller was in most cases double-logging messages.  This removes the double logging, and the reqLogger parameter, simplifying the function and leaving the logging in the context of the reconcile loop itself, for readability.

This also adds context to log messages, and in some cases, removes fmt.Sprint() usage by converting additional data to the key/value parameters to log.Logger (the reqLogger.Info() function).

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>